[app/send-receive] factor out address-autocomplete as a separate component
